feat: implement CropScreen for image cropping functionality
